pyproject.toml
[project]
name = "video"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.13.5",
    "apify>=3.3.3",
    "chromadb>=1.5.8",
    "fastapi>=0.136.1",
    "jinja2>=3.1.6",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-tavily>=0.2.18",
    "langgraph>=1.1.10",
    "loguru>=0.7.3",
    "mcp>=1.27.0",
    "openai>=2.33.0",
    "pdfkit>=1.0.0",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "reportlab>=4.4.10",
    "requests>=2.33.1",
    "tavily-python>=0.7.24",
    "typing-extensions>=4.15.0",
    "uvicorn>=0.46.0",
]
